models/ifltpp/requirements.txt
#python=3.7
pytorch=1.8.1